+++ global.d.ts
@@ -3,4 +3,9 @@ declare module '*.scss' {
   export default content
 }
 
+declare module '*.css' {
+  const content: string
+  export default content
+}
+
 declare module 'rollup-plugin-serve'