+++ types/index.d.ts
@@ -1,4 +1,4 @@
-import { DotLottiePlayer } from './elements/DotLottiePlayer';
+import DotLottiePlayer from './elements/DotLottiePlayer';
 export default DotLottiePlayer;
 export { PlayMode, PlayerEvents, PlayerState } from './enums';
 export declare const tagName = "dotlottie-player";