+++ types/templates/controls.d.ts
@@ -1,2 +1,2 @@
-import type { DotLottiePlayer } from '../elements/DotLottiePlayer';
+import type DotLottiePlayer from '../elements/DotLottiePlayer';
 export default function renderControls(this: DotLottiePlayer): void;